+++ packages/reactive_ble_mobile/android/src/main/kotlin/com/signify/hue/flutterreactiveble/ble/BleClient.kt
@@ -74,4 +74,6 @@ interface BleClient {
         characteristic: UUID,
         value: ByteArray
     )
+
+    fun isDeviceConnected(deviceId: String): Boolean
 }